add favoritesDataManager
